Explore SIMD-generic concepts
The whole discussion about regularity leads to the question of the
concepts around SIMD. std::simd<T> is a generalization of T in that the
trivial (width == 1) case reduces back to T. A more general system of
types needs an according set of concepts, otherwise we can't describe
the system clearly. This set of concepts might actually be very useful
in checking consistency of the whole design.

ChangeLog:

	* constexpr_tests.c++: Test simd_integral, simd_floating_point,
	simd_regular, simd_equality_comparable, and
	simd_equality_comparable_with on all std::simd types.
	Extra tests for simd_generic::integral and std::simd_regular.
	* fwddecl.h: Declare bool overloads of mask reductions.
	* simd_meta.h (__valid_simd, __valid_mask, __boolean_reducable)
	(__simd_weakly_equality_comparable_with)
	(__simd_partially_ordered_with)
	(__simd_comparison_common_type_with, simd_integral)
	(simd_signed_integral, simd_unsigned_integral)
	(simd_floating_point, simd_arithmetic, simd_equality_comparable)
	(simd_equality_comparable_with, simd_totally_ordered)
	(simd_totally_ordered_with, simd_regular): New concepts.
	(simd_generic::integral, signed_integral, unsigned_integral)
	(simd_generic::floating_point, arithmetic, regularity): New
	concepts.